refactor(cognition): move recency-rendering tests into recency.rs (pure code-motion)
Relocated the five self-contained recency tests (humanize_result_content, render_act_for_recall,
summarize_args_for_recency, bound_recency_result) from mod.rs's test mod into recency.rs's own
`#[cfg(test)] mod tests` — pure-function unit tests needing no executor/faculty harness. Dropped
the now-unused ContextBudget and recency-fn imports from mod.rs's test imports. Behavior-identical.
mod.rs 1527 → 1355 lines; recency.rs 215 → 387 lines. act_observe tests green: 28 passed, 0 failed.
